test: keep Linear recovery coverage scoped
